third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

CONSOLE MESSAGE: line 1: [object DOMException]
CONSOLE MESSAGE: line 1: PASS
CONSOLE MESSAGE: line 1: PASS
CONSOLE MESSAGE: line 1: PASS
CONSOLE MESSAGE: line 1: [object DOMException]
 
